+++ Infrastructure/SQLScripts/SqlLiteWorkScripts.sql
@@ -2,6 +2,7 @@
 
 
 
+select * from SearchEvents where id = 1
 
 
 
@@ -10,21 +11,7 @@
 
 
 
-
--- 
-SELECT [Id]
-      ,[Subject] 
-      ,[CourseId]
-      ,[Deleted]
-  FROM [SonarTasks]
-  
- --
-
- select * from SonarProcess;
- 
-
- 
- 
+-- Запро выбоки задач пользователей 
  SELECT 
     st.Id AS TaskId,
     st.Title AS TaskTitle,
@@ -55,42 +42,4 @@ JOIN SearchRequests sreq ON se.RequestId = sreq.Id
 JOIN SearchAnnouncements sa ON sreq.Id = sa.Id
 JOIN Animals a ON sa.AnimalId = a.Id
 LEFT JOIN Users u ON st.AssignedTo = u.Id;
- 
- -- select * from SearchEvents
 
-
-
-
-
-
-
-
-----------
-
-DROP TABLE SonarTasks
-DROP TABLE SonarProcess
-drop table SearchLeader
-
-
--- <<<<<<<<<<  22.01.2025 настраиваю таблицу SonarTasks
-/*  
-CREATE TABLE [SonarTasks] (
-  [Id] INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL
-, [Subject] text NULL
-, [SonarProcessId] bigint NOT NULL
-, [Deleted] bigint NOT NULL
-, CONSTRAINT [FK_SonarTasks_0_0] FOREIGN KEY ([SonarProcessId]) REFERENCES [SonarProcess] ([Id]) ON DELETE CASCADE ON UPDATE NO ACTION
-);
-CREATE INDEX [SonarTasks_IX_SonarProcessId] ON [SonarTasks] ([SonarProcessId] ASC);
-*/
--- переносим данные таблицы
-/*
-INSERT INTO SonarTasks (id, Subject, SonarProcessId, Deleted)
-SELECT id, Subject, CourseId SonarProcessId, Deleted
-FROM SonarTasksOld;
-*/
---drop table SonarTasksOld
-  
-  
-  
-
