+++ SonarEventSevice.sln
@@ -31,6 +31,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Services.UnitTests", "Infra
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "SQLScripts", "SQLScripts", "{EA498AD1-0310-48A7-A250-372B9B68C051}"
 	ProjectSection(SolutionItems) = preProject
+		Infrastructure\SQLScripts\SqlLiteCretaeTables.sql = Infrastructure\SQLScripts\SqlLiteCretaeTables.sql
 		Infrastructure\SQLScripts\SqlLiteWorkScripts.sql = Infrastructure\SQLScripts\SqlLiteWorkScripts.sql
 	EndProjectSection
 EndProject
